refactor(card-form): limit saveCard input to card update data
